+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for "Blowing fuse".
 - Tests for "Container terminal".
+- Tests for "Fax machine".
 
 ## [1.4.0] - 2022-03-03
 ### Added